+++ src/test/files/tables.ts
@@ -2,6 +2,29 @@ export let testTables = [
 	{
 		input: `\
 |Left header|Center header |Right header|Default header|
+|||||
+|Line:1 Column:A|Line:1 Column:B|Line:1 Column:C|Line:1 Column:D|
+|Line:2 Col:A|Line:2 Col:B|Line:2 Col:C|Line:2 Col:D|
+|L:3 C:A|L:3 C:B|L:3 C:C|L:3 C:D|`,
+		expected: `\
+| Left header     | Center header   | Right header    | Default header  |
+|-----------------|-----------------|-----------------|-----------------|
+| Line:1 Column:A | Line:1 Column:B | Line:1 Column:C | Line:1 Column:D |
+| Line:2 Col:A    | Line:2 Col:B    | Line:2 Col:C    | Line:2 Col:D    |
+| L:3 C:A         | L:3 C:B         | L:3 C:C         | L:3 C:D         |`,
+		settings: {
+			spacePadding: 1,
+			keepFirstAndLastPipes: true,
+			defaultTableJustification: 'Left',
+			markdownGrammarScopes: ['markdown'],
+			limitLastColumnPadding: false,
+			removeColonsIfSameAsDefault: false,
+			globalColumnSizes: 'Same column size'
+		}
+	},
+	{
+		input: `\
+|Left header|Center header |Right header|Default header|
 |:-|::|-:|-|
 |Line:1 Column:A|Line:1 Column:B|Line:1 Column:C|Line:1 Column:D|
 |Line:2 Col:A|Line:2 Col:B|Line:2 Col:C|Line:2 Col:D|
@@ -596,9 +619,9 @@ export let testTables = [
 | Is source control used?    | \`NO\`      |       |
 | Are changes peer reviewed? | \`PARTIAL\` |       |
 
-| Topic                          | Status        |
-|:-------------------------------|:--------------|
-| Is Iot used?                   | \`NO\`          |`,
+| Topic                 | Status                 |
+|:----------------------|:-----------------------|
+| Is Iot used?          | \`NO\`                   |`,
 		settings: {
 			spacePadding: 1,
 			keepFirstAndLastPipes: true,