+++ src/test/suite/extension.test.ts
@@ -3,11 +3,8 @@ import * as assert from 'assert';
 // as well as import your extension to test it
 import * as vscode from 'vscode';
 import { MarkdownTableFormatterSettings } from '../../interfaces';
-import { MDTable } from '../../MDTable';
-import { tableRegex } from '../../regex';
+import { discoverMaxColumnSizes, discoverMaxTableSizes, pad, tablesIn } from '../../utils';
 import { testTables } from '../files/tables';
-import XRegExp = require('xregexp');
-import { setupMaster } from 'cluster';
 
 suite('Extension Test Suite', () => {
 
@@ -39,72 +36,20 @@ suite('Extension Test Suite', () => {
 			let uri = vscode.Uri.parse('test-table:' + i);
 			return vscode.workspace.openTextDocument(uri).then(doc => {
 				let tables = tablesIn(doc, doc.validateRange(new vscode.Range(0, 0, doc.lineCount + 1, 0)));
-				let formattedTables;
 				if (testSettings.globalColumnSizes === 'Same column size') {
-					let maxSize = tables.map(table => {
-						return table.columnSizes;
-					}).reduce((p, c) => {
-						let length = p.length > c.length ? p.length : c.length;
-						let previousBigger = p.length > c.length;
-						let result = p.length > c.length ? p : c;
-						for (let index = 0; index < length; index++) {
-							if (previousBigger) {
-								if (c[index] > p[index]) {
-									result[index] = c[index];
-								}
-							} else {
-								if (p[index] > c[index]) {
-									result[index] = p[index];
-								}
-							}
-						}
-						return result;
-					});
-					formattedTables = tables.map(table => {
+					let maxSize = discoverMaxColumnSizes(tables);
+					tables.forEach(table => {
 						table.columnSizes = maxSize;
-						return table.formatted(testTable.settings || settings);
-					}).join('\n\n');
+					});
 				} if (testSettings.globalColumnSizes === 'Same table size') {
-					let maxTableWidth = tables.map(table => {
-						return table.columnSizes;
-					}).map(sizeArray => {
-						return { size: sumArray(sizeArray), columns: sizeArray.length };
-					}).reduce((p, c) => {
-						return p.size > c.size ? p : c;
-					});
-					let tableSizes = tables.map(table => {
-						return table.columnSizes;
-					}).map((sizeArray, i) => {
-						console.log(`sizeArray:${sizeArray}`);
-						if (sumArray(sizeArray) !== maxTableWidth.size) {
-							let currentTableColumns = sizeArray.length;
-							return sizeArray.map(size => {
-								// let extraWidth = ((maxTableWidth.columns * settings.spacePadding) + maxTableWidth.columns + 1);
-								// let extraWithCurrentTable = ((currentTableColumns * settings.spacePadding) + currentTableColumns + 1);
-								// let widthAjust = (extraWithCurrentTable - extraWidth) / currentTableColumns;
-								// console.log(`extraWidth:${extraWidth}, extraWithCurrentTable:${extraWithCurrentTable}, currentTableColumns:${currentTableColumns}, widthAjust:${widthAjust}`);
-								console.log(`size:${size}, sumArray(sizeArray):${sumArray(sizeArray)}, maxTableWidth.size:${maxTableWidth.size}`);
-								let notAjusted = ((size / sumArray(sizeArray)) * maxTableWidth.size);
-								let ajustment = (maxTableWidth.columns - currentTableColumns);
-								let noRound = notAjusted + ajustment;
-								let result = Math.round(noRound);
-								console.log(`notAjusted:${notAjusted}, ajustment:${ajustment}, noRound:${noRound}, result:${result}`);
-								return result;
-							});
-						} else {
-							return sizeArray;
-						}
-					});
-					formattedTables = tables.map((table, i) => {
+					let tableSizes = discoverMaxTableSizes(tables, testSettings.spacePadding);
+					tables.forEach((table, i) => {
 						table.columnSizes = tableSizes[i];
-						return table.formatted(testTable.settings || settings);
-					}).join('\n\n');
-
-				} else {
-					formattedTables = tables.map(table => {
-						return table.formatted(testTable.settings || settings);
-					}).join('\n\n');
+					});
 				}
+				let formattedTables = tables.map(table => {
+					return table.formatted(testTable.settings || settings);
+				}).join('\n\n');
 				assert.equal(formattedTables, testTable.expected);
 			});
 		});
@@ -117,28 +62,3 @@ const testTablesProvider = new class implements vscode.TextDocumentContentProvid
 		return result;
 	}
 };
-
-function sumArray(array: number[]): number {
-	return array.reduce((p, c) => p + c);
-}
-
-function pad(text: string, columns: number): string {
-	return (' '.repeat(columns) + text).slice(-columns);
-}
-
-function tablesIn(document: vscode.TextDocument, range: vscode.Range): MDTable[] {
-	var items: MDTable[] = [];
-
-	const text = document.getText(range);
-	var pos = 0, match;
-	while ((match = XRegExp.exec(text, tableRegex, pos, false))) {
-		pos = match.index + match[0].length;
-		let offset = document.offsetAt(range.start);
-		let start = document.positionAt(offset + match.index);
-		let text = match[0].replace(/^\n+|\n+$/g, '');
-		let end = document.positionAt(offset + match.index + text.length);
-		let table = new MDTable(offset, start, end, text);
-		items.push(table);
-	}
-	return items;
-}